chore(suite-common): Trading: export HandleExchangeRequestThunkProps type
